fix eslint issue
